Use 'long' for execution counts in coverage output
It is possible in large projects for tests to trigger the execution of
some lines or functions more than 2^31-1 times (the signed int limit).
In these cases, combining merging reports will produce nonsense
outputs, perhaps even negative counts, or simply fail to parse the JSON
gcov output if a single test file triggers the problem.

+++ tools/test/CoverageOutputGenerator/java/com/google/devtools/coverageoutputgenerator/GcovJsonParser.java
@@ -104,7 +104,7 @@ public class GcovJsonParser {
     int start_line;
     String name;
     int blocks_executed;
-    int execution_count;
+    long execution_count;
     String demangled_name;
     int start_column;
     int end_line;
@@ -112,7 +112,7 @@ public class GcovJsonParser {
 
   static class GcovJsonLine {
     GcovJsonBranch[] branches;
-    int count;
+    long count;
     int line_number;
     boolean unexecuted_block;
     String function_name;
@@ -120,7 +120,7 @@ public class GcovJsonParser {
 
   static class GcovJsonBranch {
     boolean fallthrough;
-    int count;
+    long count;
 
     @SerializedName("throw")
     boolean _throw;
